DynamicalSystemsBase.jl

DynamicalSystemsBase — Module

DynamicalSystemsBase.jl

A Julia package that defines the DynamicalSystem interface and many concrete implementations used in the DynamicalSystems.jl ecosystem.

To install it, run import Pkg; Pkg.add("DynamicalSystemsBase").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

The DynamicalSystem API

DynamicalSystemsBase.DynamicalSystem — Type

DynamicalSystem

DynamicalSystem is an abstract supertype encompassing all concrete implementations of what counts as a "dynamical system" in the DynamicalSystems.jl library.

All concrete implementations of DynamicalSystem can be iteratively evolved in time via the step! function. Hence, most library functions that evolve the system will mutate its current state and/or parameters. See the documentation online for implications this has on e.g., parallelization.

DynamicalSystem is further separated into two abstract types: ContinuousTimeDynamicalSystem, DiscreteTimeDynamicalSystem. The simplest and most common concrete implementations of a DynamicalSystem are DeterministicIteratedMap or CoupledODEs.

Description

Note

The documentation of DynamicalSystem follows chapter 1 of Nonlinear Dynamics, Datseris & Parlitz, Springer 2022.

A ds::DynamicalSystem representes a flow Φ in a state space. It mainly encapsulates three things:

A state, typically referred to as u, with initial value u0. The space that u occupies is the state space of ds and the length of u is the dimension of ds (and of the state space).

A dynamic rule, typically referred to as f, that dictates how the state evolves/changes with time when calling the step! function. f is a standard Julia function, see below.

A parameter container p that parameterizes f. p can be anything, but in general it is recommended to be a type-stable mutable container.

In sort, any set of quantities that change in time can be considered a dynamical system, however the concrete subtypes of DynamicalSystem are much more specific in their scope. Concrete subtypes typically also contain more information than the above 3 items.

In this scope dynamical systems have a known dynamic rule f defined as a standard Julia function. Observed or measured data from a dynamical system are represented using AbstractDataset and are finite. Such data are obtained from the trajectory function or from an experimental measurement of a dynamical system with an unknown dynamic rule.

Construction instructions on f and u

Most of the concrete implementations of DynamicalSystem, with the exception of ArbitrarySteppable, have two ways of implementing the dynamic rule f, and as a consequence the type of the state u. The distinction is done on whether f is defined as an in-place (iip) function or out-of-place (oop) function.

oop : f must be in the form f(u, p, t) -> out which means that given a state u::SVector{<:Real} and some parameter container p it returns the output of f as an SVector{<:Real} (static vector).

iip : f must be in the form f!(out, u, p, t) which means that given a state u::AbstractArray{<:Real} and some parameter container p, it writes in-place the output of f in out::AbstractArray{<:Real}. The function must return nothing as a final statement.

t stands for current time in both cases. iip is suggested for systems with high dimension and oop for small. The break-even point is between 10 to 100 dimensions but should be benchmarked on a case-by-case basis as it depends on the complexity of f.

Autonomous vs non-autonomous systems

Whether the dynamical system is autonomous (f doesn't depend on time) or not, it is still necessary to include t as an argument to f. Some algorithms utilize this information, some do not, but we prefer to keep a consistent interface either way. You can also convert any system to autonomous by making time an additional variable. If the system is non-autonomous, its effective dimensionality is dimension(ds)+1.

API

The API that the interface of DynamicalSystem employs is the functions listed below. Once a concrete instance of a subtype of DynamicalSystem is obtained, it can quieried or altered with the following functions.

The main use of a concrete dynamical system instance is to provide it to downstream functions such as lyapunovspectrum from ChaosTools.jl or basins_of_attraction from Attractors.jl. A typical user will likely not utilize directly the following API, unless when developing new algorithm implementations that use dynamical systems.

API - information

ds(t) with ds an instance of DynamicalSystem: return the state of ds at time t. For continuous time systems this interpolates and extrapolates, while for discrete time systems it only works if t is the current time.

current_state

initial_state

current_parameters

initial_parameters

isdeterministic

isdiscretetime

dynamic_rule

current_time

initial_time

isinplace

API - alter status

reinit!

set_state!

set_parameter!

set_parameters!

DynamicalSystemsBase.current_state — Function

current_state(ds::DynamicalSystem) → u

Return the current state of ds. This state is mutated when ds is mutated.

DynamicalSystemsBase.initial_state — Function

initial_state(ds::DynamicalSystem) → u0

Return the initial state of ds. This state is never mutated and is set when initializing ds.

DynamicalSystemsBase.current_parameters — Function

current_parameters(ds::DynamicalSystem) → p

Return the current parameter container of ds. This is mutated in functions that need to evolve ds across a parameter range.

The following convenience syntax is also possible:

current_parameters(ds::DynamicalSystem, index)

which will give the specific parameter from the container at the given index (which works for arrays, dictionaries, or composite types if index is Symbol).

DynamicalSystemsBase.initial_parameters — Function

initial_parameters(ds::DynamicalSystem) → p0

Return the initial parameter container of ds. This is never mutated and is set when initializing ds.

DynamicalSystemsBase.isdeterministic — Function

isdeterministic(ds::DynamicalSystem) → true/false

Return true if ds is deterministic, i.e., the dynamic rule contains no randomness. This is information deduced from the type of ds.

DynamicalSystemsBase.isdiscretetime — Function

isdiscretetime(ds::DynamicalSystem) → true/false

Return true if ds operates in discrete time, or false if it is in continuous time. This is information deduced from the type of ds.

DynamicalSystemsBase.dynamic_rule — Function

dynamic_rule(ds::DynamicalSystem) → f

Return the dynamic rule of ds. This is never mutated and is set when initializing ds.

DynamicalSystemsBase.current_time — Function

current_time(ds::DynamicalSystem) → t

Return the current time that ds is at. This is mutated when ds is evolved.

DynamicalSystemsBase.initial_time — Function

initial_time(ds::DynamicalSystem) → t0

Return the initial time defined for ds. This is never mutated and is set when initializing ds.

SciMLBase.isinplace — Method

isinplace(ds::DynamicalSystem) → true/false

Return true if the dynamic rule of ds is in-place, i.e., a function mutating the state in place. If true, the state is typically Array, if false, the state is typically SVector. A front-end user will most likely not care about this information, but a developer may care.

SciMLBase.reinit! — Method

reinit!(ds::DynamicalSystem, u = initial_state(ds); kwargs...) → ds

Reset the status of ds, so that it is as if it has be just initialized with initial state u. Practically every function of the ecosystem that evolves ds first calls this function on it. Besides the new initial state u, you can also configure the keywords t0 = initial_time(ds) and p = current_parameters(ds).

Note the default settings: the state and time are the initial, but the parameters are the current.

The special method reinit!(ds, ::Nothing; kwargs...) is also available, which does nothing and leaves the system as is. This is so that downstream functions that call reinit! can still be used without resetting the system but rather continuing from its exact current state.

DynamicalSystemsBase.set_state! — Function

set_state!(ds::DynamicalSystem, u)

Set the state of ds to u, which must match dimensionality with that of ds. Also ensure that the change is notified to whatever integration protocol is used.

DynamicalSystemsBase.set_parameter! — Function

set_parameter!(ds::DynamicalSystem, index, value)

Change a parameter of ds given the index it has in the parameter container and the value to set it to. This function works for both array/dictionary containers as well as composite types. In the latter case index needs to be a Symbol.

DynamicalSystemsBase.set_parameters! — Function

set_parameters!(ds::DynamicalSystem, p = initial_parameters(ds))

Set the parameter values in the current_parameters(ds) to match p. This is done as an in-place overwrite by looping over the keys of p. Hence the keys of p must be a subset of the keys of current_parameters(ds).

Time evolution

SciMLBase.step! — Method

step!(ds::DiscreteTimeDynamicalSystem [, dt::Integer]) → ds

Evolve the discrete time dynamical system for 1 or dt steps.

step!(ds::ContinuousTimeDynamicalSystem, [, dt::Real [, stop_at_tdt]]) → ds

Evolve the continuous time dynamical system for one integration step.

Alternative, if a dt is given, then progress the integration until there is a temporal difference ≥ dt (so, step at least for dt time).

When true is passed to the optional third argument, the integration advances for exactly dt time.

DynamicalSystemsBase.trajectory — Function

trajectory(ds::DynamicalSystem, T [, u0]; kwargs...) → X, t

Return a dataset X that will contain the trajectory of the system ds, after evolving it for total time T. u0 is the state given given to reinit! prior to time evolution and defaults to initial_state(ds).

See Dataset for info on how to use X. The returned time vector is t = (t0+Ttr):Δt:(t0+Ttr+T).

Keyword arguments

Δt: Time step of value output. For discrete time systems it must be an integer. Defaults to 0.1 for continuous and 1 for discrete time systems.

Ttr = 0: Transient time to evolve the initial state before starting saving states.

t0 = initial_time(ds): Starting time.

save_idxs::AbstractVector{Int}: Which variables to output in X (by default all).

StateSpaceSets.Dataset — Type

Dataset{D, T} <: AbstractDataset{D,T}

A dedicated interface for datasets. It contains equally-sized datapoints of length D, represented by SVector{D, T}. These data are a standard Julia Vector{SVector}, and can be obtained with vec(dataset).

When indexed with 1 index, a dataset is like a vector of datapoints. When indexed with 2 indices it behaves like a matrix that has each of the columns be the timeseries of each of the variables.

Dataset also supports most sensible operations like append!, push!, hcat, eachrow, among others, and when iterated over, it iterates over its contained points.

Description of indexing

In the following let i, j be integers, typeof(data) <: AbstractDataset and v1, v2 be <: AbstractVector{Int} (v1, v2 could also be ranges, and for massive performance benefits make v2 an SVector{X, Int}).

data[i] == data[i, :] gives the ith datapoint (returns an SVector)

data[v1] == data[v1, :], returns a Dataset with the points in those indices.

data[:, j] gives the jth variable timeseries, as Vector

data[v1, v2], data[:, v2] returns a Dataset with the appropriate entries (first indices being "time"/point index, while second being variables)

data[i, j] value of the jth variable, at the ith timepoint

Use Matrix(dataset) or Dataset(matrix) to convert. It is assumed that each column of the matrix is one variable. If you have various timeseries vectors x, y, z, ... pass them like Dataset(x, y, z, ...). You can use columns(dataset) to obtain the reverse, i.e. all columns of the dataset in a tuple.

DeterministicIteratedMap

DynamicalSystemsBase.DeterministicIteratedMap — Type

DeterministicIteratedMap <: DynamicalSystem
DeterministicIteratedMap(f, u0, p = nothing; t0 = 0)

A deterministic discrete time dynamical system defined by an iterated map as follows:

\[\vec{u}_{n+1} = \vec{f}(\vec{u}_n, p, n)\]

An alias for DeterministicIteratedMap is DiscreteDynamicalSystem.

Optionally configure the parameter container p and initial time t0.

For construction instructions regarding f, u0 see DynamicalSystem.

CoupledODEs

DynamicalSystemsBase.CoupledODEs — Type

CoupledODEs <: ContinuousTimeDynamicalSystem
CoupledODEs(f, u0 [, p]; diffeq, t0 = 0.0)

A deterministic continuous time dynamical system defined by a set of coupled ordinary differential equations as follows:

\[\frac{d\vec{u}}{dt} = \vec{f}(\vec{u}, p, t)\]

An alias for CoupledODE is ContinuousDynamicalSystem.

Optionally provide the parameter container p and initial time as keyword t0.

For construction instructions regarding f, u0 see DynamicalSystem.

DifferentialEquations.jl keyword arguments and interfacing

The ODEs are evolved via the solvers of DifferentialEquations.jl. When initializing a CoupledODEs, you can specify the solver that will integrate f in time, along with any other integration options, using the diffeq keyword. For example you could use diffeq = (abstol = 1e-9, reltol = 1e-9). If you want to specify a solver, do so by using the keyword alg, e.g.: diffeq = (alg = Tsit5(), maxiters = 100000). This requires you to have been first using OrdinaryDiffEq to access the solvers. The default diffeq is:

(alg = Tsit5(stagelimiter! = triviallimiter!, steplimiter! = triviallimiter!, thread = static(false)), abstol = 1.0e-6, reltol = 1.0e-6)

diffeq keywords can also include callback for event handling , however the majority of downstream functions in DynamicalSystems.jl assume that f is differentiable.

The convenience constructor CoupledODEs(prob::ODEProblem, diffeq) is also available.

Dev note: CoupledODEs is a light wrapper of ODEIntegrator from DifferentialEquations.jl. The integrator is available as the field integ, and the ODEProblem is integ.sol.prob. The convenience syntax ODEProblem(ds::CoupledODEs, tspan = (t0, Inf)) is available.

StroboscopicMap

DynamicalSystemsBase.StroboscopicMap — Type

StroboscopicMap <: DiscreteTimeDynamicalSystem
StroboscopicMap(ds::CoupledODEs, T::Real) → smap

A discrete time dynamical system that produces iterations of a time-dependent (non-autonomous) CoupledODEs system exactly over a period T. This is known as a stroboscopic map. The second signature creates a CoupledODEs and calls the first (with t0 measured in continuous time).

As this system is in discrete time, current_time and initial_time are integers. The initial time is always 0, because current_time counts elapsed periods. Call these functions on the field .ds of StroboscopicMap to obtain the corresponding continuous time. In contrast, reinit! expects t0 in continuous time.

The convenience constructor

StroboscopicMap(T::Real, f, u0, p = nothing; diffeq, t0 = 0) → smap

is also provided.

See also PoincareMap.

PoincareMap

DynamicalSystemsBase.PoincareMap — Type

PoincareMap <: DiscreteTimeDynamicalSystem
PoincareMap(ds::CoupledODEs, plane; kwargs...) → pmap

A discrete time dynamical system that produces iterations over the Poincaré map[DatserisParlitz2022] of the given continuous time ds. This map is defined as the sequence of points on the Poincaré surface of section, which is defined by the plane argument.

See also StroboscopicMap, poincaresos, produce_orbitdiagram.

Keyword arguments

direction = -1: Only crossings with sign(direction) are considered to belong to the surface of section. Positive direction means going from less than $b$ to greater than $b$.

u0 = nothing: Specify an initial state.

rootkw = (xrtol = 1e-6, atol = 1e-6): A NamedTuple of keyword arguments passed to find_zero from Roots.jl.

Tmax = 1e3: The argument Tmax exists so that the integrator can terminate instead of being evolved for infinite time, to avoid cases where iteration would continue forever for ill-defined hyperplanes or for convergence to fixed points, where the trajectory would never cross again the hyperplane. If during one step! the system has been evolved for more than Tmax, then step!(pmap) will terminate and error.

Description

The Poincaré surface of section is defined as sequential transversal crossings a trajectory has with any arbitrary manifold, but here the manifold must be a hyperplane. PoincareMap iterates over the crossings of the section.

If the state of ds is $\mathbf{u} = (u_1, \ldots, u_D)$ then the equation defining a hyperplane is

\[a_1u_1 + \dots + a_Du_D = \mathbf{a}\cdot\mathbf{u}=b\]

where $\mathbf{a}, b$ are the parameters of the hyperplane.

In code, plane can be either:

A Tuple{Int, <: Real}, like (j, r): the plane is defined as when the jth variable of the system equals the value r.

A vector of length D+1. The first D elements of the vector correspond to $\mathbf{a}$ while the last element is $b$.

PoincareMap uses ds, higher order interpolation from DifferentialEquations.jl, and root finding from Roots.jl, to create a high accuracy estimate of the section.

PoincareMap follows the DynamicalSystem interface with the following adjustments:

dimension(pmap) == dimension(ds), even though the Poincaré map is effectively 1 dimension less.

Like StroboscopicMap time is discrete and counts the iterations on the surface of section. initial_time is always 0 and current_time is current iteration number.

A new function current_crossing_time returns the real time corresponding to the latest crossing of the hyperplane, which is what the current_state(ds) corresponds to as well.

For the special case of plane being a Tuple{Int, <:Real}, a special reinit! method is allowed with input state of length D-1 instead of D, i.e., a reduced state already on the hyperplane that is then converted into the D dimensional state.

Example

using DynamicalSystemsBase
ds = Systems.rikitake(zeros(3); μ = 0.47, α = 1.0)
pmap = poincaremap(ds, (3, 0.0))
step!(pmap)
next_state_on_psos = current_state(pmap)

DynamicalSystemsBase.current_crossing_time — Function

current_crossing_time(pmap::PoincareMap) → tcross

Return the time of the latest crossing of the Poincare section.

DynamicalSystemsBase.poincaresos — Function

poincaresos(A::AbstractDataset, plane; kwargs...) → P::Dataset

Calculate the Poincaré surface of section of the given dataset with the given plane by performing linear interpolation betweeen points that sandwich the hyperplane.

Argument plane and keywords direction, warning, save_idxs are the same as in the method below.

poincaresos(ds::CoupledODEs, plane, T = 1000.0; kwargs...) → P::Dataset

Return the iterations of ds on the Poincaré surface of section with the plane, by evolving ds up to a total of T. Return a Dataset of the points that are on the surface of section.

This function initializes a PoincareMap and steps it until its current_crossing_time exceeds T. You can also use trajectory with PoincareMap to get a sequence of N::Int points instead.

The keywords Ttr, save_idxs act as in trajectory. See PoincareMap for plane and all other keywords.

ProjectedDynamicalSystem

DynamicalSystemsBase.ProjectedDynamicalSystem — Type

ProjectedDynamicalSystem <: DynamicalSystem
ProjectedDynamicalSystem(ds::DynamicalSystem, projection, complete_state)

A dynamical system that represents a projection of an existing ds on a (projected) space.

The projection defines the projected space. If projection isa AbstractVector{Int}, then the projected space is simply the variable indices that projection contains. Otherwise, projection can be an arbitrary function that given the state of the original system ds, returns the state in the projected space. In this case the projected space can be equal, or even higher-dimensional, than the original.

complete_state produces the state for the original system from the projected state. complete_state can always be a function that given the projected state returns a state in the original space. However, if projection isa AbstractVector{Int}, then complete_state can also be a vector that contains the values of the remaining variables of the system, i.e., those not contained in the projected space. In this case the projected space needs to be lower-dimensional than the original.

Notice that ProjectedDynamicalSystem does not require an invertible projection, complete_state is only used during reinit!. ProjectedDynamicalSystem is in fact a rather trivial wrapper of ds which steps it as normal in the original state space and only projects as a last step, e.g., during current_state.

Examples

Case 1: project 5-dimensional system to its last two dimensions.

ds = Systems.lorenz96(5)
projection = [4, 5]
complete_state = [0.0, 0.0, 0.0] # completed state just in the plane of last two dimensions
pds = projected_integrator(ds, projection, complete_state)
reinit!(pds, [0.2, 0.4])
step!(pds)
get_state(pds)

Case 2: custom projection to general functions of state. julia ds = Systems.lorenz96(5) projection(u) = [sum(u), sqrt(u[1]^2 + u[2]^2)] complete_state(y) = repeat(y[1]/5, 5) pds = # same as in above example...`

ParallelDynamicalSystem

DynamicalSystemsBase.ParallelDynamicalSystem — Type

ParallelDynamicalSystem <: DynamicalSystem
ParallelDynamicalSystem(ds::DynamicalSystem, states::Vector{<:AbstractArray})

A struct that evolves several states of a given dynamical system in parallel at exactly the same times. Useful when wanting to evolve several different trajectories of the same system while ensuring that they share parameters and time vector.

This struct follows the DynamicalSystem interface with the following adjustments:

The function current_state is called as current_state(pds, i::Int = 1) which returns the ith state. Same for initial_state.

Similarly, set_state! obtains a second argument i::Int = 1 to set the i-th state.

current_states and initial_states can be used to get all parallel states.

reinit! takes in a vector of states (like states) for u.

DynamicalSystemsBase.initial_states — Function

initial_states(pds::ParallelDynamicalSystem)

Return an iterator over the initial parallel states of pds.

DynamicalSystemsBase.current_states — Function

current_states(pds::ParallelDynamicalSystem)

Return an iterator over the parallel states of pds.

TangentDynamicalSystem

DynamicalSystemsBase.CoreDynamicalSystem — Type

CoreDynamicalSystem

Union type meaning either DeterministicIteratedMap or CoupledODEs, which are the core systems whose dynamic rule f is known analytically.

This type is used for deciding whether a creation of a TangentDynamicalSystem is possible or not.

DynamicalSystemsBase.TangentDynamicalSystem — Type

TangentDynamicalSystem <: DynamicalSystem
TangentDynamicalSystem(ds::CoreDynamicalSystem; kwargs...)

A dynamical system that bundles the evolution of ds (which must be an CoreDynamicalSystem) and k deviation vectors that are evolved according to the dynamics in the tangent space (also called linearized dynamics or the tangent dynamics).

The state of ds must be an AbstractVector for TangentDynamicalSystem.

TangentDynamicalSystem follows the DynamicalSystem interface with the following adjustments:

reinit! takes an additional keyword Q0 (with same default as below)

The additional functions current_deviations and set_deviations! are provided for the deviation vectors.

Keyword arguments

k or Q0: Q0 represents the initial deviation vectors (each column = 1 vector). If k::Int is given, a matrix Q0 is created with the first k columns of the identity matrix. Otherwise Q0 can be given directly as a matrix. It must hold that size(Q, 1) == dimension(ds). You can use orthonormal for random orthonormal vectors. By default k = dimension(ds) is used.

u0 = current_state(ds): Starting state.

J and J0: See section "Jacobian" below.

Description

Let $u$ be the state of ds, and $y$ a deviation (or perturbation) vector. These two are evolved in parallel according to

\[\begin{array}{rcl} \frac{d\vec{x}}{dt} &=& f(\vec{x}) \\ \frac{dY}{dt} &=& J_f(\vec{x}) \cdot Y \end{array} \quad \mathrm{or}\quad \begin{array}{rcl} \vec{x}_{n+1} &=& f(\vec{x}_n) \\ Y_{n+1} &=& J_f(\vec{x}_n) \cdot Y_n. \end{array}\]

for continuous or discrete time respectively. Here $f$ is the dynamic_rule(ds) and $J_f$ is the Jacobian of $f$.

Jacobian

The keyword J provides the Jacobian function. It must be a Julia function in the same form as f, the dynamic_rule. Specifically, J(u, p, n) -> M::SMatrix for the out-of-place version or J(M, u, p, n) for the in-place version acting in-place on M. in both cases M is a matrix whose columns are the deviation vectors.

By default J = nothing. In this case J is constructed automatically using the module ForwardDiff, hence its limitations also apply here. Even though ForwardDiff is very fast, depending on your exact system you might gain significant speed-up by providing a hand-coded Jacobian and so it is recommended. Additionally, automatic and in-place Jacobians cannot be time dependent.

The keyword J0 allows you to pass an initialized Jacobian matrix J0. This is useful for large in-place systems where only a few components of the Jacobian change during the time evolution. J0 can be a sparse or any other matrix type. If not given, a matrix of zeros is used. J0 is ignored for out of place systems.

DynamicalSystemsBase.current_deviations — Function

current_deviations(tands::TangentDynamicalSystem)

Return the deviation vectors of tands as a matrix with each column a vector.

DynamicalSystemsBase.set_deviations! — Function

set_deviations!(tands::TangentDynamicalSystem, Q)

Set the deviation vectors of tands to be Q, a matrix with each column a vector.

StateSpaceSets.orthonormal — Function

orthonormal([T,] D, k) -> ws

Return a matrix ws with k columns, each being an D-dimensional orthonormal vector. T is the return type and can be either SMatrix or Matrix

returns either SMatrix{D, k} if D*k < 100, otherwise Dxk Matrix{Float64}

DatserisParlitz2022Datseris & Parlitz 2022, Nonlinear Dynamics: A Concise Introduction Interlaced with Code, Springer Nature, Undergrad. Lect. Notes In Physics